# apps/api/openapi/libif-api.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/register": {
      "post": {
        "operationId": "register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          }
        },
        "summary": "Register a reader account and start a session.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/sign-in": {
      "post": {
        "operationId": "signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          }
        },
        "summary": "Sign in with email and password.",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/auth/password-reset-requests": {
      "post": {
        "operationId": "requestPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMessageDto"
                }
              }
            }
          }
        },
        "summary": "Request a password reset link if an account exists.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/password-resets": {
      "post": {
        "operationId": "resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMessageDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          }
        },
        "summary": "Reset a password with a valid reset token.",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/access/documents/{documentId}/decision": {
      "get": {
        "operationId": "getDecision",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessDecisionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Check access decision for a document by reader or staff role.",
        "tags": [
          "Access"
        ]
      }
    },
    "/api/access/documents/{documentId}/manifest": {
      "get": {
        "operationId": "getManifest",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtectedDocumentManifestDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get protected document manifest metadata for reader canvas viewing.",
        "tags": [
          "Access"
        ]
      }
    },
    "/api/access/documents/{documentId}/search": {
      "get": {
        "operationId": "searchDocumentText",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "minLength": 2,
              "maxLength": 100,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPageSearchResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search authorized document text and return up to five matching pages.",
        "tags": [
          "Access"
        ]
      }
    },
    "/api/access/documents/{documentId}/pages/{pageNumber}": {
      "get": {
        "operationId": "getPage",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get an authorized, server-watermarked page image.",
        "tags": [
          "Access"
        ]
      }
    },
    "/api/access/documents/{documentId}/view-token": {
      "post": {
        "operationId": "createViewToken",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtectedDocumentUrlDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtain a protected viewing URL token for staff or legacy viewer.",
        "tags": [
          "Access"
        ]
      }
    },
    "/api/access/documents/{documentId}/download-token": {
      "post": {
        "operationId": "createDownloadToken",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtectedDocumentUrlDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtain a protected download URL token (staff only).",
        "tags": [
          "Access"
        ]
      }
    },
    "/api/access/documents/{documentId}/stream": {
      "get": {
        "operationId": "streamDocument",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Stream protected document PDF content (staff internal).",
        "tags": [
          "Access"
        ]
      }
    },
    "/api/access/documents/{documentId}/file": {
      "get": {
        "operationId": "downloadDocument",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Download protected document PDF file (staff internal).",
        "tags": [
          "Access"
        ]
      }
    },
...
  },
  "info": {
    "title": "LIBIF API",
    "description": "OpenAPI contract for LIBIF web and API integration.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "reader@example.edu"
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "example": "correct horse battery staple"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SessionUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "dev-librarian"
          },
          "email": {
            "type": "string",
            "example": "librarian@libif.local"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "LIBRARIAN",
              "READER"
            ]
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "catalogue:read",
                "reader:library:read",
                "admin:books:read",
                "admin:books:write",
                "admin:taxonomy:manage",
                "admin:users:manage"
              ]
            }
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "permissions"
        ]
      },
      "SessionDto": {
        "type": "object",
        "properties": {
          "authenticated": {
            "type": "boolean",
            "example": true
          },
          "user": {
            "$ref": "#/components/schemas/SessionUserDto"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "catalogue:read",
                "reader:library:read",
                "admin:books:read",
                "admin:books:write",
                "admin:taxonomy:manage",
                "admin:users:manage"
              ]
            }
          },
          "strategy": {
            "type": "string",
            "enum": [
              "persistent-cookie",
              "development-header"
            ]
          }
        },
        "required": [
          "authenticated",
          "permissions",
          "strategy"
        ]
      },
      "AuthErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "AUTHENTICATION_REQUIRED"
          },
          "message": {
            "type": "string",
            "example": "Authentication is required."
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "example": {}
          },
          "traceId": {
            "type": "string",
            "example": "request-trace-id"
          },
          "status": {
            "type": "number",
            "example": 403
          }
        },
        "required": [
          "code",
          "message",
          "fieldErrors",
          "traceId",
          "status"
        ]
      },
      "SignInRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "reader@example.edu"
          },
          "password": {
            "type": "string",
            "example": "correct horse battery staple"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthMessageDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "If an account exists for that email, a reset link has been sent."
          }
        },
        "required": [
          "message"
        ]
      },
      "PasswordResetRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "reader@example.edu"
          }
        },
        "required": [
          "email"
        ]
      },
      "PasswordResetDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "reset-token-from-email"
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "example": "new correct horse battery staple"
          }
        },
        "required": [
          "token",
          "password"
        ]
      },
...
      "AccessDecisionDto": {
        "type": "object",
        "properties": {
          "allowed": {
            "type": "boolean",
            "example": true,
            "description": "Whether the current user is eligible to view/download this document"
          },
          "documentId": {
            "type": "string",
            "example": "cm123456789"
          },
          "userRole": {
            "type": "string",
            "example": "READER"
          },
          "documentStatus": {
            "type": "string",
            "example": "PUBLISHED",
            "description": "Current lifecycle status of the document (reader-safe subset)",
            "enum": [
              "DRAFT",
              "PENDING_PROCESSING",
              "PROCESSING",
              "PENDING_APPROVAL",
              "CORRECTION_REQUIRED",
              "PUBLISHED",
              "REJECTED"
            ]
          },
          "reason": {
            "type": "string",
            "example": "Document is published and available."
          }
        },
        "required": [
          "allowed",
          "documentId",
          "userRole"
        ]
      },
      "ProtectedPageDescriptorDto": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "number",
            "minimum": 1
          },
          "width": {
            "type": "number",
            "minimum": 1
          },
          "height": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "pageNumber",
          "width",
          "height"
        ]
      },
      "ProtectedDocumentManifestDto": {
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string"
          },
          "pageCount": {
            "type": "number",
            "minimum": 1
          },
          "minZoom": {
            "type": "number",
            "example": 0.5
          },
          "maxZoom": {
            "type": "number",
            "example": 2
          },
          "pages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProtectedPageDescriptorDto"
            }
          }
        },
        "required": [
          "documentId",
          "pageCount",
          "minZoom",
          "maxZoom",
          "pages"
        ]
      },
      "DocumentPageSearchResultDto": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "number"
          },
          "matchCount": {
            "type": "number"
          },
          "snippet": {
            "type": "string"
          }
        },
        "required": [
          "pageNumber",
          "matchCount"
        ]
      },
      "DocumentPageSearchResponseDto": {
        "type": "object",
        "properties": {
          "available": {
            "type": "boolean"
          },
          "query": {
            "type": "string"
          },
          "results": {
            "maxItems": 5,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentPageSearchResultDto"
            }
          },
          "totalMatches": {
            "type": "number"
          },
          "totalPagesWithMatches": {
            "type": "number"
          },
          "truncated": {
            "type": "boolean"
          }
        },
        "required": [
          "available",
          "query",
          "results",
          "totalMatches",
          "totalPagesWithMatches",
          "truncated"
        ]
      },
      "ProtectedDocumentUrlDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "access_token_abc123"
          },
          "expiresAt": {
            "type": "string",
            "example": "2026-07-21T11:45:00.000Z"
          },
          "url": {
            "type": "string",
            "example": "/api/access/stream?token=access_token_abc123"
          }
        },
        "required": [
          "token",
          "expiresAt",
          "url"
        ]
      },
...
    }
  }
}